Extract CSP account lookup concern
