+++ dpc-portal/app/controllers/concerns/csp_existing_account.rb
@@ -1,69 +1,34 @@
 # frozen_string_literal: true
 
-# Handles merging existing emails and accounts
+# Handles routing CSP user responses to the correct flow
 module CspExistingAccount
   extend ActiveSupport::Concern
+  include CspAccountLookup
 
   private
 
   def handle_csp_user_response(csp_user, auth)
-    return render_add_email(csp_user, auth) if csp_user && !email_match?(csp_user, auth)
+    csp_user ? handle_existing_user(csp_user, auth) : handle_unlinked_account(auth)
+  end
 
-    if csp_user.nil?
-      orig_csp_user = existing_account(auth)
-      return render_link_account(primary_email(auth), orig_csp_user.csp.name) if orig_csp_user.present?
-    end
+  def handle_existing_user(csp_user, auth)
+    return render_add_email(csp_user, auth) unless email_match?(csp_user, auth)
 
     check_csp_session
     sync_and_redirect(csp_user, auth)
   end
 
+  def handle_unlinked_account(auth)
+    orig_csp_user = existing_account(auth)
+    return render_link_account(primary_email(auth), orig_csp_user.csp.name) if orig_csp_user.present?
+
+    check_csp_session
+    csp_user = current_user&.csp_user_for(auth.provider)
+    sync_and_redirect(csp_user, auth)
+  end
+
   def email_match?(csp_user, auth)
     csp_user.user_emails.empty? || csp_user.user_emails.map(&:email).include?(primary_email(auth))
-  end
-
-  def name_match?(user, auth)
-    info = auth.extra.raw_info
-    user.given_name.casecmp?(info.given_name) && user.family_name.casecmp?(info.family_name)
-  end
-
-  def existing_account(auth)
-    csp_users = matching_csp_users(auth)
-    return nil if csp_users.empty?
-
-    csp_users.first
-  end
-
-  def matching_csp_users(auth)
-    emails = other_csp_emails(auth)
-    return [] if emails.empty?
-
-    matching, mismatched = emails.map(&:csp_user).uniq.partition { |c| name_match?(c.user, auth) }
-    log_name_mismatch(mismatched) if mismatched.any?
-    validate_unique_match(matching, emails)
-    matching
-  end
-
-  def other_csp_emails(auth)
-    UserEmail.includes(csp_user: %i[user csp])
-             .where(email: [primary_email(auth), *all_emails(auth)])
-             .reject { |e| e.csp_user.csp.name == auth.provider.to_s }
-  end
-
-  def log_name_mismatch(csp_users)
-    csp_users.each do |csp_user|
-      Rails.logger.info(['Email match found but name does not match',
-                         { actionContext: LoggingConstants::ActionContext::Authentication,
-                           actionType: LoggingConstants::ActionType::NameMismatch,
-                           csp: csp_user.csp.name }])
-    end
-  end
-
-  def validate_unique_match(csp_users, emails)
-    return unless csp_users.many?
-
-    matching_emails = emails.map(&:email).uniq.join(',')
-    raise CspUtils::MultiUserMatchError, "too many matching users | #{matching_emails}"
   end
 
   def render_add_email(csp_user, auth)
